GRHB-64: + useHandleClickOutside hook, * wrap image inside button
