+++ frontend/src/components/common/header/styles.module.scss
@@ -28,3 +28,8 @@
 .popup {
   margin-left: auto;
 }
+
+.button {
+  background-color: transparent;
+  border: none;
+}
