+++ frontend/src/hooks/hooks.ts
@@ -1,7 +1,8 @@
-export { useCallback, useEffect, useMemo, useState } from 'react';
+export { useCallback, useEffect, useMemo, useState, useRef } from 'react';
 export { useLocation, useNavigate, useParams } from 'react-router-dom';
 
 export { useAppDispatch } from './use-app-dispatch/use-app-dispatch.hook';
 export { useAppForm } from './use-app-form/use-app-form.hook';
 export { useAppSelector } from './use-app-selector/use-app-selector.hook';
 export { useFormControl } from './use-form-control/use-form-control.hook';
+export { useHandleClickOutside } from './use-handle-click-outside/use-handle-click-outside.hook';
